Starting Portfolio README
This is the start of my portfolio, as you can see, I only have one project within the portfolio as of now. I will be adding more projects as I go along in the freecodecamp.com tutorials. And at the bottom it shows the coding for the navigation bar as well as the backgrounds and colors used for text and different pages. Again, if there is any insight at all, please feel free to share.